requirements.txt
apache-airflow==2.7.0
psycopg2-binary==2.9.7
requests==2.31.0
pandas==2.0.3
sqlalchemy==1.4.49